requirements.txt
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
google-auth==2.49.2
google-auth-oauthlib==1.3.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mod_wsgi==5.0.2
oauthlib==3.3.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyMySQL==1.1.2
python-dotenv==1.2.2
requests==2.33.1
requests-oauthlib==2.0.0
SQLAlchemy==2.0.49
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8